int8_t are not printed as numbers

+++ Tracking/TrkFitter/TrkGaussianSumFilter/test/testMergeComponents.cxx
@@ -9,8 +9,8 @@
 
 using namespace GSFUtils;
 namespace {
-constexpr int16_t n =72;
-constexpr std::array<Component1D,n> input = {
+constexpr int16_t n = 72;
+constexpr std::array<Component1D, n> input = {
   { { -4.66462e-06, 1.06618e-11, 9.37928e+10, 0.00608503 },
     { -2.08263e-05, 7.533e-11, 1.32749e+10, 0.0274963 },
     { -4.15487e-05, 4.79975e-11, 2.08344e+10, 0.0591673 },
@@ -99,8 +99,9 @@ main()
   }
   std::vector<std::pair<int8_t, int8_t>> mergeOrder =
     findMerges(components.buffer(), n, 12);
-  for (const auto& i : mergeOrder){
-    std::cout << "[" << i.first << ", " << i.second << "]" << '\n';
+  for (const auto& i : mergeOrder) {
+    std::cout << "[" << static_cast<int>(i.first) << ", "
+              << static_cast<int>(i.second) << "]" << '\n';
   }
   return 0;
 }
